refactor(practice): move measureRangeToTicks to its own module
Exporting a non-component from PracticeViewPlugin.tsx breaks React Fast
Refresh (react-refresh/only-export-components lint error). Move to a
standalone measureRangeToTicks.ts and import from there.
