+++ frontend/plugins/practice-view-plugin/measureRangeToTicks.test.ts
@@ -13,7 +13,7 @@
  */
 
 import { describe, it, expect } from 'vitest';
-import { measureRangeToTicks } from './PracticeViewPlugin';
+import { measureRangeToTicks } from './measureRangeToTicks';
 
 // ---------------------------------------------------------------------------
 // measureRangeToTicks
